# packages/api-spec/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zenodotus API",
    "description": "Local server API for intelligent LLM-powered tab grouping",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "operationId": "healthCheck",
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Server is running",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/group": {
      "post": {
        "operationId": "groupTabs",
        "summary": "Group browser tabs using LLM",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tabs"
                ],
                "properties": {
                  "tabs": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "tabId",
                        "windowId",
                        "url",
                        "title"
                      ],
                      "properties": {
                        "tabId": {
                          "type": "integer",
                          "description": "Chrome tab ID"
                        },
                        "windowId": {
                          "type": "integer",
                          "description": "Chrome window ID"
                        },
                        "url": {
                          "type": "string",
                          "description": "Tab URL"
                        },
                        "title": {
                          "type": "string",
                          "description": "Tab title"
                        },
                        "description": {
                          "type": "string",
                          "description": "Page meta description"
                        }
                      }
                    }
                  },
                  "existingGroups": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "groupId",
                        "name",
                        "tabIds"
                      ],
                      "properties": {
                        "groupId": {
                          "type": "integer",
                          "description": "Chrome tab group ID"
                        },
                        "name": {
                          "type": "string",
                          "description": "Group name"
                        },
                        "tabIds": {
                          "type": "array",
                          "items": {
                            "type": "integer"
                          },
                          "description": "Tab IDs currently in this group"
                        }
                      }
                    }
                  },
                  "prompt": {
                    "type": "string",
                    "description": "User's custom grouping strategy prompt"
                  },
                  "model": {
                    "type": "string",
                    "description": "LLM model to use (e.g. sonnet, opus, haiku)"
                  },
                  "thinking": {
                    "type": "boolean",
                    "description": "Enable extended thinking"
                  },
                  "provider": {
                    "type": "string",
                    "description": "LLM provider (claude-code or codex)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tab grouping result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "groups"
                  ],
                  "properties": {
                    "groups": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "tabIds"
                        ],
                        "properties": {
                          "groupId": {
                            "type": "integer",
                            "description": "ID of an existing group. Omit to create a new group."
                          },
                          "name": {
                            "type": "string",
                            "description": "Name for the group. Required when creating a new group."
                          },
                          "tabIds": {
                            "type": "array",
                            "items": {
                              "type": "integer"
                            },
                            "description": "Tab IDs to assign to this group"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
      "TabInfo": {
        "type": "object",
        "required": [
          "tabId",
          "windowId",
          "url",
          "title"
        ],
        "properties": {
          "tabId": {
            "type": "integer",
            "description": "Chrome tab ID"
          },
          "windowId": {
            "type": "integer",
            "description": "Chrome window ID"
          },
          "url": {
            "type": "string",
            "description": "Tab URL"
          },
          "title": {
            "type": "string",
            "description": "Tab title"
          },
          "description": {
            "type": "string",
            "description": "Page meta description"
          }
        }
      },
      "ExistingGroup": {
        "type": "object",
        "required": [
          "groupId",
          "name",
          "tabIds"
        ],
        "properties": {
          "groupId": {
            "type": "integer",
            "description": "Chrome tab group ID"
          },
          "name": {
            "type": "string",
            "description": "Group name"
          },
          "tabIds": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "Tab IDs currently in this group"
          }
        }
      },
      "GroupRequest": {
        "type": "object",
        "required": [
          "tabs"
        ],
        "properties": {
          "tabs": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "tabId",
                "windowId",
                "url",
                "title"
              ],
              "properties": {
                "tabId": {
                  "type": "integer",
                  "description": "Chrome tab ID"
                },
                "windowId": {
                  "type": "integer",
                  "description": "Chrome window ID"
                },
                "url": {
                  "type": "string",
                  "description": "Tab URL"
                },
                "title": {
                  "type": "string",
                  "description": "Tab title"
                },
                "description": {
                  "type": "string",
                  "description": "Page meta description"
                }
              }
            }
          },
          "existingGroups": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "groupId",
                "name",
                "tabIds"
              ],
              "properties": {
                "groupId": {
                  "type": "integer",
                  "description": "Chrome tab group ID"
                },
                "name": {
                  "type": "string",
                  "description": "Group name"
                },
                "tabIds": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  },
                  "description": "Tab IDs currently in this group"
                }
              }
            }
          },
          "prompt": {
            "type": "string",
            "description": "User's custom grouping strategy prompt"
          },
          "model": {
            "type": "string",
            "description": "LLM model to use (e.g. sonnet, opus, haiku)"
          },
          "thinking": {
            "type": "boolean",
            "description": "Enable extended thinking"
          },
          "provider": {
            "type": "string",
            "description": "LLM provider (claude-code or codex)"
          }
        }
      },
      "GroupAssignment": {
        "type": "object",
        "required": [
          "tabIds"
        ],
        "properties": {
          "groupId": {
            "type": "integer",
            "description": "ID of an existing group. Omit to create a new group."
          },
          "name": {
            "type": "string",
            "description": "Name for the group. Required when creating a new group."
          },
          "tabIds": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "Tab IDs to assign to this group"
          }
        }
      },
      "GroupResponse": {
        "type": "object",
        "required": [
          "groups"
        ],
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "tabIds"
              ],
              "properties": {
                "groupId": {
                  "type": "integer",
                  "description": "ID of an existing group. Omit to create a new group."
                },
                "name": {
                  "type": "string",
                  "description": "Name for the group. Required when creating a new group."
                },
                "tabIds": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  },
                  "description": "Tab IDs to assign to this group"
                }
              }
            }
          }
        }
      }
    }
  }
}